Added dictionary
Did a big rewrite of all the code, hopefully it works. I made it so once
you create an item (with the item script) you just have to go into the
Dictionary object in the heirarchy, edit the dictionary script, add the
object, make a public GameObject (prefabName) then add your new object
to it in the inspector. This may not make any sense because I'm
incredibly drunk. Good luck.